requirements.txt
plotly
pandas
numpy
joblib
biopython
gunicorn
django
xgboost